+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Improvements
 
 - Add support to read/add/remove embedded layer styles in gpkg (#263)
+- Add `gridsize` parameter to most spatial operations (#261)
 - Improve performance of makevalid and isvalid (#258)
 - Improve performance of spatial operations when only one batch is used (#271)
 - Add checks that `output_path` must not be equal to the/an `input_path` for geo